Storage rework (#13)
This changes how app bundles are stored. They can now be stored in one of three different repositories (Azure Storage, S3 and S3-compatible, and a local folder). Clients now upload the full bundle to the node, which puts it in storage.

This contains other changes too: allows stores bundles' hashes in addition to signatures, and hashes can be validated even if there's no signature.